Fix: 검색창 컴포넌트 생성 후 전시목록/전시리뷰목록/메이트글목록 페이지에 적용
