+++ src/components/exhibitionList/Filters.tsx
@@ -1,10 +1,12 @@
 import React from "react";
 import styled from "styled-components";
 import { theme } from "../../styles/theme";
-import { FilterType } from "../../types/exhbList";
 import Button from "../atom/Button";
+import SearchInput from "../atom/SearchInput";
 import Selectbox from "../atom/Selectbox";
 import { PagesState } from "../Pagination";
+import { alertPreparing } from "../../utils/alerts";
+import { FilterType } from "../../types/exhbList";
 
 // 전시글 목록 상단 필터 컴포넌트_박예선_23.02.08
 const Filters = (props: FiltersType) => {
@@ -68,7 +70,7 @@ const Filters = (props: FiltersType) => {
           placeholder="검색"
           onClick={() => alert("준비 중인 기능입니다.")}
           onKeyDown={(e) => {
-            if (e.key === "Enter") alert("준비 중인 기능입니다.");
+            if (e.key === "Enter") alertPreparing();
           }}
         />
       </div>
@@ -135,19 +137,5 @@ const FiltersContainer = styled.div`
         margin-right: 10px;
       }
     }
-  }
-`;
-
-const SearchInput = styled.input`
-  width: 277px;
-  height: 36px;
-  padding: 0 42px 0 20px;
-  border: 1px solid ${(props) => props.theme.colors.greys40};
-  background: url("data:image/svg+xml,%3Csvg width='24' height='24' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Ccircle cx='11' cy='11' r='6' stroke='%239C9C9C' stroke-width='2'/%3E%3Cpath d='M16.2071 14.7929L15.5 14.0858L14.0858 15.5L14.7929 16.2071L16.2071 14.7929ZM18.2929 19.7071C18.6834 20.0976 19.3166 20.0976 19.7071 19.7071C20.0976 19.3166 20.0976 18.6834 19.7071 18.2929L18.2929 19.7071ZM14.7929 16.2071L18.2929 19.7071L19.7071 18.2929L16.2071 14.7929L14.7929 16.2071Z' fill='%239C9C9C'/%3E%3C/svg%3E%0A")
-    no-repeat;
-  background-position: right 14px center;
-  font-size: 14px;
-  &::placeholder {
-    color: ${(props) => props.theme.colors.greys60};
   }
 `;
