+++ src/pages/BlogReviewList.tsx
@@ -1,14 +1,16 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import styled from "styled-components";
 import { useQuery } from "react-query";
+import styled from "styled-components";
 import { theme } from "../styles/theme";
-import Selectbox from "../components/atom/Selectbox";
-import ReviewCardList from "../components/blogReviewList/ReviewCardList";
 import Button from "../components/atom/Button";
-import blogReviewApis from "../apis/blogReviewApis";
+import Selectbox from "../components/atom/Selectbox";
+import SearchInput from "../components/atom/SearchInput";
+import ReviewCardList from "../components/blogReviewList/ReviewCardList";
 import Pagination from "../components/Pagination";
+import { alertPreparing } from "../utils/alerts";
 import { BlogReviewResponse } from "../types/blogReview";
+import blogReviewApis from "../apis/blogReviewApis";
 
 const BlogReviewList = () => {
   const navigate = useNavigate();
@@ -62,12 +64,13 @@ const BlogReviewList = () => {
           />
         </Flex>
         <Flex style={{ gap: "10px" }}>
-          <InputContainer>
-            <Input onChange={handleInputLength} />
-            <InputLegnth inputlength={inputLength}>
-              {inputLength}/60
-            </InputLegnth>
-          </InputContainer>
+          <SearchInput
+            placeholder="검색"
+            onClick={() => {}}
+            onKeyDown={(e) => {
+              if (e.key === "Enter") alertPreparing();
+            }}
+          />
           <Button
             onClick={() => navigate("/blogreview-write")}
             style={{ padding: "8px 20px" }}
@@ -123,32 +126,4 @@ const Flex = styled.div`
   display: flex;
   align-items: center;
   gap: 4px;
-`;
-
-const InputContainer = styled.div`
-  position: relative;
-`;
-
-const Input = styled.input`
-  width: 277px;
-  height: 36px;
-  margin: 10px auto;
-  padding-left: 10px;
-  border: 1px solid ${theme.colors.greys40};
-  border-radius: 30px;
-`;
-
-interface InputProps {
-  inputlength: number;
-}
-
-const InputLegnth = styled.div<InputProps>`
-  position: absolute;
-  top: 50%;
-  right: 10px;
-  transform: translateY(-50%);
-  color: ${(props) =>
-    props.inputlength > 60 ? theme.colors.error : theme.colors.greys60};
-  font-weight: 500;
-  font-size: 14px;
 `;
